Refactor authentication to align with overall project structure
