+++ src/main/kotlin/data/datasource/mapper/sesssion/SessionColumns.kt
@@ -1,0 +1,8 @@
+package org.baghdad.data.datasource.mapper.sesssion
+
+object SessionColumns {
+    const val id = 0
+    const val userId = 1
+    const val token = 3
+    const val loginTime = 4
+}
