+++ src/main/kotlin/data/local/SessionDataSource.kt
@@ -1,7 +1,7 @@
 package org.baghdad.data.local
 
+import SessionEntity
 import org.baghdad.data.datasource.DataSource
-import org.baghdad.logic.entities.authentication.SessionEntity
 
 class SessionDataSource(
     dataSource: DataSource<SessionEntity>
