+++ src/main/kotlin/logic/model/exceptions/AuthenticationExceptions.kt
@@ -1,0 +1,5 @@
+package org.baghdad.logic.model.exceptions
+
+class InvalidPasswordException(message: String) : IllegalArgumentException(message)
+class UserCanNotBeFoundException(message: String) : Exception(message)
+class SessionEndedException(message: String) : Exception(message)
